requirements.txt
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-generativeai>=0.3.0
supabase>=2.4.0,<3
postgrest>=0.14.8
pytz
requests
beautifulsoup4
youtube-transcript-api>=0.6.0
numpy
pandas
scipy
nltk==3.8.1
textblob==0.17.1
opencv-python-headless==4.10.0.84
Pillow>=10.4.0
imagehash>=4.3.1
pytesseract==0.3.10
language-tool-python
unidecode
ultralytics
sentence-transformers
torch
scikit-learn
isodate>=0.6.1
